requirements.txt
beautifulsoup4>=4.11.1
requests>=2.28.1
langchain-google-genai>=0.1.0
langchain-core>=0.3.0
python-dotenv>=0.20.0
flask>=2.0.0
